+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### Version 0.4.1
+
+- Added "hours" digits to timer
+- Fixed a bug where the rare drop sound would continue playing forever
+
 ### Version 0.4
 - Added rare drop alerts
 - Merged IronwoodNotify into the tracker